Slight change to not create a blank data file

+++ python/Ganga/Core/GangaRepository/GangaRepositoryXML.py
@@ -89,10 +89,6 @@ def safe_save(fn, _obj, to_file, ignore_subs=''):
             if not os.path.exists(dirname):
                 os.makedirs(dirname)
 
-            # Create a blank file if it's not already there (append just in case)
-            if not os.path.exists(fn):
-                open(fn, 'a').close()
-
             # Prepare new data file
             new_name = fn + '.new'
             with open(new_name, "w") as tmpfile:
@@ -100,7 +96,11 @@ def safe_save(fn, _obj, to_file, ignore_subs=''):
 
             # everything ready so create new data file and backup old one
             if os.path.exists(new_name):
-                os.rename(fn, fn + "~")
+
+                # Do we have an old one to backup?
+                if os.path.exists(fn):
+                    os.rename(fn, fn + "~")
+
                 os.rename(new_name, fn)
 
         except (IOError, OSError) as err:
